+++ core/src/main/resources/org/apache/commons/validator/javascript/validateMask-compress.js
@@ -1,0 +1,33 @@
+function validateMask(_1){
+var _2=true;
+var _3=null;
+var i=0;
+var _5=new Array();
+var _6=eval("new "+jcv_retrieveFormName(_1)+"_mask()");
+for(var x in _6){
+if(!jcv_verifyArrayElement(x,_6[x])){
+continue;
+}
+var _8=_1[_6[x][0]];
+if(!jcv_isFieldPresent(_8)){
+continue;
+}
+if((_8.type=="hidden"||_8.type=="text"||_8.type=="textarea"||_8.type=="file")&&(_8.value.length>0)){
+if(!jcv_matchPattern(_8.value,_6[x][2]("mask"))){
+if(i==0){
+_3=_8;
+}
+_5[i++]=_6[x][1];
+_2=false;
+}
+}
+}
+if(_5.length>0){
+jcv_handleErrors(_5,_3);
+}
+return _2;
+}
+function jcv_matchPattern(_9,_a){
+return _a.exec(_9);
+}
+